feat(Yuicy): 修复一些潜在bug，添加一些新功能 - 添加2D灯光能力 - Window添加光标修改接口，支持无边框窗口(自定义窗口关闭、最小化等图标，支持窗口拖动) - 添加一些方便开发的接口 - 消除一些编译警告

+++ Yuicy/src/Yuicy/Renderer/OrthographicCameraController.cpp
@@ -87,7 +87,7 @@ namespace Yuicy {
 		if (e.GetHeight() == 0.0f)  // 窗口最小化时会触发高度为0的事件
 			return false;
 
-		m_AspectRatio = e.GetWidth() / e.GetHeight();
+		m_AspectRatio = static_cast<float>(e.GetWidth()) / static_cast<float>(e.GetHeight());
 		CalculateView();
 		return false;
 	}
